Organized tests and added non-working test for iteration runner
